Added MCP73831 Charger
Deceided to design around using a 3.7 Li-Poly flat pack battery.  To
make it more user friendly (ie don't need to fumble getting the battery
out) I have added a MCP73831 Charger.  Currently set to 500 mA charge.
This should be fine for any Li-Poly battery that has a maximum voltage
of 4.2 and a minimum capacity of 500 mAh.  Batteries with smaller
capacity may be damaged by the higher charge current as most batteries
are a 1C charge rate.